\name{makeExample}
\alias{makeExample}
\title{Make ARF example data directory structure.}
\description{Make ARF example data directory structure.}
\usage{
makeExample(path)
}
\arguments{
  \item{path}{Optional path to create experiment directories in.}
}
\value{Saves and loads the experiment}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\keyword{ models }
